updating per convention, to sync with 0.17.0 standards
