+++ .env-ci
@@ -1,0 +1,4 @@
+CELLULOID_SPECS_LOG_STRATEGY=stderr
+CELLULOID_SPECS_LOG_LEVEL=3
+CELLULOID_SPECS_LOG_FILE=log/ci.log
+CELLULOID_SPECS_LOG_SYNC=false
