+++ .env-dev
@@ -1,0 +1,4 @@
+CELLULOID_SPECS_LOG_STRATEGY=single
+CELLULOID_SPECS_LOG_FILE=log/test.log
+CELLULOID_SPECS_LOG_LEVEL=0
+CELLULOID_SPECS_LOG_SYNC=true
