+++ Gemfile
@@ -1,8 +1,7 @@
 source 'http://rubygems.org'
 
 gem 'coveralls', require: false
-gem 'celluloid', github: 'celluloid/celluloid', branch: 'master'
+gem 'celluloid', github: 'celluloid/celluloid', branch: '0.17.0-prerelease', submodules: true
 gem 'transpec', github: 'yujinakayama/transpec', tag: 'v3.1.0'
 
-# Specify your gem's dependencies in celluloid-zmq.gemspec
 gemspec
