+++ lib/celluloid/zmq/deprecate.rb
@@ -1,0 +1,9 @@
+module Celluloid
+  module ZMQ
+    ReadableSocket = Socket::Readable
+    WritableSocket = Socket::Writable
+
+    RepSocket = Socket::Rep
+    ReqSocket = Socket::Req
+  end
+end
