+++ lib/celluloid/zmq/mailbox.rb
@@ -1,7 +1,7 @@
 module Celluloid
   module ZMQ
     # Replacement mailbox for Celluloid::ZMQ actors
-    class Mailbox < Celluloid::EventedMailbox
+    class Mailbox < Celluloid::Mailbox::Evented
       def initialize
         super(Reactor)
       end
